+++ src/Rascal.Tests/ResultExtensionsTests.cs
@@ -8,7 +8,7 @@ public class ResultExtensionsTests
         var r = Ok(Ok(2));
         var x = r.Unnest();
 
-        x.HasValue.ShouldBeTrue();
+        x.IsOk.ShouldBeTrue();
         x.value.ShouldBe(2);
     }
 
@@ -18,7 +18,7 @@ public class ResultExtensionsTests
         var x = "uwu";
         var r = x.NotNull("error");
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe("uwu");
     }
 
@@ -28,7 +28,7 @@ public class ResultExtensionsTests
         var x = null as string;
         var r = x.NotNull("error");
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -38,7 +38,7 @@ public class ResultExtensionsTests
         var x = 2 as int?;
         var r = x.NotNull("error");
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe(2);
     }
 
@@ -48,7 +48,7 @@ public class ResultExtensionsTests
         var x = null as int?;
         var r = x.NotNull("error");
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -77,7 +77,7 @@ public class ResultExtensionsTests
 
         var result = xs.Sequence();
         
-        result.HasValue.ShouldBeTrue();
+        result.IsOk.ShouldBeTrue();
         result.value.ShouldBe([1, 2, 3, 4, 5]);
     }
 
@@ -88,7 +88,7 @@ public class ResultExtensionsTests
 
         var result = xs.Sequence();
 
-        result.HasValue.ShouldBeFalse();
+        result.IsOk.ShouldBeFalse();
         result.error?.Message.ShouldBe("error 1");
     }
 
@@ -99,7 +99,7 @@ public class ResultExtensionsTests
 
         var result = xs.Sequence();
 
-        result.HasValue.ShouldBeTrue();
+        result.IsOk.ShouldBeTrue();
         result.value.ShouldBeEmpty();
     }
 
@@ -109,7 +109,7 @@ public class ResultExtensionsTests
         var dict = new Dictionary<string, int>() { ["a"] = 2 };
         var r = dict.TryGetValueR("a", "error");
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe(2);
     }
 
@@ -119,7 +119,7 @@ public class ResultExtensionsTests
         var dict = new Dictionary<string, int>();
         var r = dict.TryGetValueR("a", "error");
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -129,7 +129,7 @@ public class ResultExtensionsTests
         var xs = new[] { 2 };
         var r = xs.Index(0, "error");
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe(2);
     }
 
@@ -139,7 +139,7 @@ public class ResultExtensionsTests
         var xs = Array.Empty<int>();
         var r = xs.Index(0, "error");
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -148,7 +148,7 @@ public class ResultExtensionsTests
     {
         var r = await Task.FromResult(2).CatchCancellation();
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe(2);
     }
 
@@ -161,7 +161,7 @@ public class ResultExtensionsTests
         await cts.CancelAsync();
         var r = await task.CatchCancellation();
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error.ShouldBeOfType<CancellationError>();
         return;