+++ src/Rascal.Tests/UtilityTests.cs
@@ -10,7 +10,7 @@ public class UtilityTests
         Result<int> x = 2;
 
         x.value.ShouldBe(2);
-        x.HasValue.ShouldBeTrue();
+        x.IsOk.ShouldBeTrue();
     }
 
     [Fact]
@@ -20,7 +20,7 @@ public class UtilityTests
         var b = Ok("uwu");
         var r = a.Combine(b);
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe((2, "uwu"));
     }
 
@@ -31,7 +31,7 @@ public class UtilityTests
         var b = Err<string>("error");
         var r = a.Combine(b);
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -42,7 +42,7 @@ public class UtilityTests
         var b = Ok("uwu");
         var r = a.Combine(b);
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -53,7 +53,7 @@ public class UtilityTests
         var b = Err<string>("error b");
         var r = a.Combine(b);
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         var error = r.error.ShouldBeOfType<AggregateError>();
         error.Errors.Select(x => x.Message).ShouldBe(["error a", "error b"]);
     }
@@ -63,7 +63,7 @@ public class UtilityTests
     {
         var r = Ok<object>("uwu").ToType<string>();
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe("uwu");
     }
 
@@ -72,7 +72,7 @@ public class UtilityTests
     {
         var r = Ok<object>("uwu").ToType<int>();
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldNotBeNull();
     }
 
@@ -81,7 +81,7 @@ public class UtilityTests
     {
         var r = Err<object>("error").ToType<int>();
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
 
@@ -90,7 +90,7 @@ public class UtilityTests
     {
         var r = Ok(2).Validate(_ => false);
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldNotBeNull();
     }
 
@@ -99,7 +99,7 @@ public class UtilityTests
     {
         var r = Ok(2).Validate(_ => false, _ => "error");
         
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error");
     }
     
@@ -108,7 +108,7 @@ public class UtilityTests
     {
         var r = Ok(2).Validate(_ => true, _ => "error");
 
-        r.HasValue.ShouldBeTrue();
+        r.IsOk.ShouldBeTrue();
         r.value.ShouldBe(2);
     }
 
@@ -117,7 +117,7 @@ public class UtilityTests
     {
         var r = Err<int>("error a").Validate(_ => false, _ => "error b");
 
-        r.HasValue.ShouldBeFalse();
+        r.IsOk.ShouldBeFalse();
         r.error?.Message.ShouldBe("error a");
     }