+++ src/Rascal/ResultDebugProxy.cs
@@ -2,9 +2,9 @@ namespace Rascal;
 
 internal sealed class ResultDebugProxy<T>(Result<T> result)
 {
-    public bool HasValue { get; } = result.HasValue;
+    public bool IsOk { get; } = result.IsOk;
 
-    public object? Value { get; } = result.HasValue
+    public object? Value { get; } = result.IsOk
         ? result.value!
         : result.Error;
 }