+++ CHANGELOG.md
@@ -1,1 +1,7 @@
 # Version `1.1.0-pre`
+
+- Rewrite and update a lot of documentation.
+
+- Rename `Result<T>.HasValue` to `IsOk`.
+
+- Add `Result<T>.IsError`.